Extract form to own component so we don't re-render the whole page on form input change
